requirements.txt
anyascii==0.3.1
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
Django==4.0.4
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-filter==21.1
django-modelcluster==5.3
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
executing==0.8.3
gunicorn==20.1.0
html5lib==1.1
idna==3.3
ipython==8.3.0
jedi==0.18.1
l18n==2021.3
matplotlib-inline==0.1.3
openpyxl==3.0.9
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
tablib==3.2.1
telepath==0.2
traitlets==5.2.0
urllib3==1.26.9
wagtail==2.16.2
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.1.0
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
